requirements.txt
beautifulsoup4==4.8.2
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-debug-toolbar==2.2
django-heroku==0.3.1
django-widget-tweaks==1.4.5
gunicorn==19.9.0
idna==2.9
newsapi-python==0.2.6
psycopg2==2.8.4
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.0.1
wincertstore==0.2
